🤖 refactor(site/src/pages): nest organization settings pages
