+++ site/src/pages/OrganizationSettingsPage/OrganizationGeneralSettingsPage/OrganizationSettingsPageView.stories.tsx
@@ -7,7 +7,8 @@ import {
 import { OrganizationSettingsPageView } from "./OrganizationSettingsPageView";
 
 const meta: Meta<typeof OrganizationSettingsPageView> = {
-	title: "pages/OrganizationSettingsPageView",
+	title:
+		"pages/OrganizationSettingsPage/OrganizationGeneralSettingsPage/OrganizationSettingsPageView",
 	component: OrganizationSettingsPageView,
 	args: {
 		organization: MockOrganization,
